+++ src/main/java/com/verdantartifice/thaumicwonders/common/misc/GuiIds.java
@@ -1,0 +1,5 @@
+package com.verdantartifice.thaumicwonders.common.misc;
+
+public class GuiIds {
+    public static final int TIMEWINDER = 1;
+}