+++ changelog.md
@@ -1,5 +1,6 @@
 v1.1.0
 ------
+* Added Timewinder, a device that lets you skip forward in time
 * The Primal Destroyer now truly hungers; keep it fed, or else
 * Increased Primal Destroyer damage to compensate